// Net.Vatprc.Uniapi.UI.Operation/src/api.d.ts
export interface components {
  schemas: {
// ...
    FlightDto: {
      id: components["schemas"]["Ulid"];
      cid: string;
      callsign: string;
      /** Format: date-time */
      last_observed_at: string;
      departure: string;
      arrival: string;
      equipment: string;
      navigation_performance: string;
      transponder: string;
      raw_route: string;
      __simplified_route: string;
      aircraft: string;
    };
// ...
    WarningMessage: {
      message_code: string;
      parameter?: string | null;
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}